requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
--index-url https://pypi.douban.com/simple/

attrs==20.3.0
    # via jsonschema
bson==0.5.10
    # via -r requirements.in
click==7.1.2
    # via
    #   -r requirements.in
    #   flask
flask==1.1.2
    # via
    #   -c requirements_fix.txt
    #   -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
itsdangerous==1.1.0
    # via flask
jinja2==2.11.2
    # via
    #   -c requirements_fix.txt
    #   flask
jsonschema==3.2.0
    # via
    #   -c requirements_fix.txt
    #   -r requirements.in
markupsafe==1.1.1
    # via jinja2
numpy==1.18.0
    # via
    #   -c requirements_fix.txt
    #   -r requirements.in
    #   pandas
pandas==1.0.4
    # via
    #   -c requirements_fix.txt
    #   -r requirements.in
pymongo==3.11.2
    # via -r requirements.in
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   -c requirements_fix.txt
    #   bson
    #   pandas
pytz==2020.5
    # via pandas
pyyaml==5.4.1
    # via -r requirements.in
six==1.15.0
    # via
    #   bson
    #   jsonschema
    #   python-dateutil
werkzeug==1.0.1
    # via
    #   -c requirements_fix.txt
    #   flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
